pyproject.toml
[project]
name = "cfd-payment"
version = "0.1.0"
description = "Analysis of UK Contracts for Difference (CfD) renewable energy subsidy payments"
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "openpyxl",
    "pandas",
    "plotly",
    "pydantic>=2.13.1",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.3",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cfd_payment"]
